Convert package manager from pnpm to npm
We did this to try and fix a module resolution issue when using this project with npm link,
but this didn't solve the issue. We left the package manager as npm because using
the same package manager can reduce conflicts when npm linking, and also because
we got cleaner results when running `npm ls react` in this project.